:sparkles: (Radius) Add radius request for IOT connections

+++ src/admin/internet/adminInternetDeleteAccess.js
@@ -1,4 +1,6 @@
 const { DatabaseSingleton } = require("../../utils/databaseSingleton");
+const { DisableRadiusIOTConnection } = require("../../utils/radius/disableRadiusIOTConnection");
+const { prepareMacAddress } = require('./adminInternetAllowAccess');
 
 /*
  * Delete a hardware request created by user
@@ -8,8 +10,15 @@ const AdminInternetDeleteAccess = (req, res) => {
     let database = DatabaseSingleton.getInstance().getDatabase();
 
     let access_id = parseInt(req.params.access_id);
-    database.query('DELETE FROM access WHERE access_id = ?', [access_id], () => {
-        res.redirect('/admin/internet/');
+    database.query('SELECT * FROM access WHERE access_id = ?', [access_id], (errors, results, fields) => {
+        if(!!results && results.length > 0) {
+            if(process.env.RADIUS == "true"){
+                DisableRadiusIOTConnection(prepareMacAddress(results[0].access_mac));
+            }
+            database.query('DELETE FROM access WHERE access_id = ?', [access_id], () => {
+                res.redirect('/admin/internet/');
+            });
+        }
     });
 }
